+++ Classes/Command/BatchTranslation.php
@@ -4,16 +4,20 @@ declare(strict_types=1);
 
 namespace ThieleUndKlose\Autotranslate\Command;
 
-use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Output\OutputInterface;
+use Doctrine\DBAL\ParameterType;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
-use TYPO3\CMS\Core\Core\Bootstrap;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+use ThieleUndKlose\Autotranslate\Domain\Model\BatchItem;
 use ThieleUndKlose\Autotranslate\Domain\Repository\BatchItemRepository;
 use ThieleUndKlose\Autotranslate\Service\BatchTranslationService;
-use TYPO3\CMS\Extbase\Persistence\Generic\PersistenceManager;
+use TYPO3\CMS\Core\Core\Bootstrap;
+use TYPO3\CMS\Core\Database\ConnectionPool;
+use TYPO3\CMS\Core\Information\Typo3Version;
+use TYPO3\CMS\Extbase\Persistence\Generic\Mapper\DataMapper;
 
 // final class BatchTranslation extends Command
 final class BatchTranslation extends Command implements LoggerAwareInterface
@@ -23,53 +27,47 @@ final class BatchTranslation extends Command implements LoggerAwareInterface
     public const ITEMS_PER_RUN_DEFAULT = 1;
 
     /**
-     * @var BatchItemRepository
-     */
-    protected $batchItemRepository;
-
-    /**
-     * @param BatchItemRepository $batchItemRepository
-     * @return void
-     */
-    public function injectBatchItemRepository(BatchItemRepository $batchItemRepository): void
-    {
-        $this->batchItemRepository = $batchItemRepository;
-    }
-
-    /**
-     * @var PersistenceManager
-     * @inject
-     */
-    protected $persistenceManager;
-
-    /**
-     * @param PersistenceManager $persistenceManager
-     * @return void
-     */
-    public function injectPersistenceManager(PersistenceManager $persistenceManager): void
-    {
-        $this->persistenceManager = $persistenceManager;
-    }
-
-    /**
      * @var BatchTranslationService
-     * @inject
      */
     protected $batchTranslationService;
 
     /**
-     * @param
-     * @return void
+     * @var ConnectionPool
      */
-    public function injectBatchTranslationService(BatchTranslationService $batchTranslationService): void
-    {
+    protected $connectionPool;
+
+    /**
+     * @var DataMapper
+     */
+    protected $dataMapper;
+
+    /**
+     * @var Typo3Version
+     */
+    protected $typo3Version;
+
+     /**
+     * @param BatchItemRepository $batchItemRepository
+     * @param ConnectionPool $connectionPool
+     * @param DataMapper $dataMapper
+     * @param Typo3Version $typo3Version
+     */
+    public function __construct(
+        BatchTranslationService $batchTranslationService,
+        ConnectionPool $connectionPool,
+        DataMapper $dataMapper,
+        Typo3Version $typo3Version
+    ) {
+        parent::__construct();
         $this->batchTranslationService = $batchTranslationService;
+        $this->connectionPool = $connectionPool;
+        $this->dataMapper = $dataMapper;
+        $this->typo3Version = $typo3Version;
     }
 
     /**
      * @return void
      */
-
     protected function configure(): void
     {
         $this
@@ -111,11 +109,10 @@ final class BatchTranslation extends Command implements LoggerAwareInterface
         // Init _cli_ backend user, needed to execute command directly from CLI
         Bootstrap::initializeBackendAuthentication();
 
-        $translationsPerRun = $input->getArgument('translationsPerRun');
-        $translationsPerRun = (int)$translationsPerRun;
+        $translationsPerRun = (int)$input->getArgument('translationsPerRun');
         $successfulTranslations = 0;
 
-        $batchItemsToRun = $this->batchItemRepository->findWaitingForRun($translationsPerRun);
+        $batchItemsToRun = $this->findWaitingForRun($translationsPerRun);
 
         if (empty($batchItemsToRun)) {
             $output->writeln('No translation to run!');
@@ -127,10 +124,9 @@ final class BatchTranslation extends Command implements LoggerAwareInterface
             if ($res === true) {
                 $successfulTranslations++;
                 $item->markAsTranslated();
+                $this->updateBatchItem($item);
             }
-            $this->batchItemRepository->update($item);
         }
-        $this->persistenceManager->persistAll();
 
         $this->logTranslationStats($successfulTranslations, $translationsPerRun);
 
@@ -139,4 +135,80 @@ final class BatchTranslation extends Command implements LoggerAwareInterface
         return Command::SUCCESS;
     }
 
+    /**
+     * find all items recursively to run
+     * TODO: filter items recursively for given page from argument
+     * @param int $limit|null
+     * @return array
+     */
+    public function findWaitingForRun(int $limit = null): array
+    {
+        $queryBuilder = $this->connectionPool->getQueryBuilderForTable('tx_autotranslate_batch_item');
+        $queryBuilder->getRestrictions()->removeAll();
+
+        $now = new \DateTime();
+        $queryBuilder
+            ->select('*')
+            ->from('tx_autotranslate_batch_item')
+            ->where(
+                // only load items where translate is gerader than translated
+                $this->typo3Version->getMajorVersion() < 11 ?
+                    $queryBuilder->expr()->orX(
+                        $queryBuilder->expr()->isNull('translated'),
+                        $queryBuilder->expr()->gt('translate', 'translated'),
+                    )
+                :
+                    $queryBuilder->expr()->or(
+                        $queryBuilder->expr()->isNull('translated'),
+                        $queryBuilder->expr()->gt('translate', 'translated'),
+                    )
+                ,
+                // only load items where error is empty
+                $queryBuilder->expr()->eq('error', $queryBuilder->createNamedParameter('')),
+                // only loaditems with next translation date in the past
+                $queryBuilder->expr()->lt('translate', $queryBuilder->createNamedParameter($now->getTimestamp())),
+                // only load active items
+                $queryBuilder->expr()->eq('hidden', $queryBuilder->createNamedParameter(false))
+            );
+
+        if ($limit) {
+            $queryBuilder->setMaxResults($limit);
+        }
+
+        if ($this->typo3Version->getMajorVersion() < 11) {
+            $statement = $queryBuilder->execute();
+        } else {
+            $statement = $queryBuilder->executeQuery();
+        }
+
+        return $this->dataMapper->map(BatchItem::class, $statement->fetchAllAssociative());
+    }
+
+    /**
+     * Update properties of proceeded batchItem
+     *
+     * @param BatchItem $batchItem
+     * @return void
+     */
+    private function updateBatchItem(BatchItem $item): void
+    {
+        $queryBuilder = $this->connectionPool->getQueryBuilderForTable('tx_autotranslate_batch_item');
+        $queryBuilder
+            ->update('tx_autotranslate_batch_item')
+            ->where(
+                $queryBuilder->expr()->eq('uid', $queryBuilder->createNamedParameter($item->getUid(), ParameterType::INTEGER))
+            )
+            ->set('error', $item->getError())
+            ->set('translate', $item->getTranslate()->getTimestamp());
+
+        if ($item->getTranslated()) {
+            $queryBuilder->set('translated', $item->getTranslated()->getTimestamp());
+        }
+
+        if ($this->typo3Version->getMajorVersion() > 10) {
+            $queryBuilder->executeStatement();
+        } else {
+            $queryBuilder->execute();
+        }
+    }
 }