+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.2.1] - 2024-10-20
+
+### Fix
+- Symfony command to perform a batch translation in Typo3 v13
+- Number of translations per run limited to the maximum defined number
+
 ## [1.2.0] - 2024-10-15 (TYPO3 v13.4 LTS Release)
 
 ### Added